Update index.rst
Found and fixed some typos in the documentation. I also have noticed that the HTML link at the end of section 7.3 is showing as HTML in the rendered documentation. I don't know yet how to fix it. This is one of my first times using Github so hope I don't make any mistakes.